refactor: refactor code to separate TaggedTemplateExpression processing
